requirements.txt
fastapi
uvicorn
openai-whisper
sentence-transformers
faiss-cpu
numpy
pydantic
python-multipart
chromadb
